Software/AIris-Core-System/requirements.txt
gradio
torch
opencv-python-headless
Pillow
transformers
groq
python-dotenv
sentence-transformers
scikit-learn
numpy